ai-tele-appointment/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
pytest
httpx
google-api-python-client
google-auth
twilio
